convert ? placeholders in where() and having() to :_#_ named placeholders
this is because PDO is really touchy about sequence numbers on ? placeholders. if we have bound values [:foo, :bar, ?, :baz], the ? placeholder is not number 1, it is number 3. as it is nigh impossible to keep track of the numbering when done out-of-order, we now do a braindead check on the where/having condition string to see if it has ? placholders, and replace them with named :_#_ placeholders, where # is the current count of the $bind_values array.